STEP 6. Folders dragon drop
