+++ altitude/static/js/frontend-app.js
@@ -37,6 +37,7 @@ export class FrontendApp {
         this.Alpine.start()
         this.bindBatchOpsDragDrop()
         this.bindPeopleDragDrop()
+        this.bindFolderDragDrop()
         this.hydrateFragments(document)
 
         return this
@@ -691,6 +692,89 @@ export class FrontendApp {
                     mergeSourceId,
                     mergeDestId,
                 })
+            },
+
+            ondropdeactivate: (event) => {
+                event.target.classList.remove("drop-active")
+                event.target.classList.remove("drop-target")
+            },
+        })
+    }
+
+    bindFolderDragDrop() {
+        interact("#rootFolderList .drag-drop").draggable({
+            inertia: true,
+            autoScroll: true,
+
+            listeners: {
+                move: dragMoveListener,
+                end: dragged,
+            },
+        })
+
+        interact("#rootFolderList .dropzone").dropzone({
+            accept: "#rootFolderList .drag-drop, #assets .drag-drop, #batchOps .drag-drop",
+            overlap: 0.2,
+
+            ondropactivate: (event) => {
+                event.target.classList.add("drop-active")
+            },
+
+            ondragenter: (event) => {
+                const draggableElement = event.relatedTarget
+                const dropzoneElement = event.target
+
+                dropzoneElement.classList.add("drop-target")
+                draggableElement.classList.add("can-drop")
+            },
+
+            ondragleave: (event) => {
+                event.target.classList.remove("drop-target")
+                event.relatedTarget.classList.remove("can-drop")
+            },
+
+            ondrop: (event) => {
+                const draggableElement = event.relatedTarget
+                const dropzoneElement = event.target
+                const movedFolderId = draggableElement.getAttribute(
+                    Const.attributes.folderId,
+                )
+                const movedAssetId = draggableElement.getAttribute(
+                    Const.attributes.assetId,
+                )
+                const isBatchMover = draggableElement.parentNode.classList.contains(
+                    "batch-mover",
+                )
+                const newParentId = dropzoneElement.getAttribute(
+                    Const.attributes.folderId,
+                )
+
+                dropzoneElement.classList.remove("drop-active")
+                dropzoneElement.classList.remove("drop-target")
+                draggableElement.classList.remove("can-drop")
+
+                if (movedFolderId) {
+                    console.debug(`Moved folder ${movedFolderId} to ${newParentId}`)
+                    this.dispatch(Const.events.folderMoved, {
+                        movedFolderId,
+                        newParentId,
+                    })
+                }
+
+                if (movedAssetId) {
+                    console.debug(`Moved asset ${movedAssetId} to ${newParentId}`)
+                    this.dispatch(Const.events.assetMoved, {
+                        assetId: movedAssetId,
+                        folderId: newParentId,
+                    })
+                }
+
+                if (isBatchMover) {
+                    console.debug(`Batch moving assets to folder ${newParentId}`)
+                    this.dispatch(Const.events.batchAssetsMoved, {
+                        folderId: newParentId,
+                    })
+                }
             },
 
             ondropdeactivate: (event) => {
