+++ altitude/views/htmx/folders.scala.html
@@ -187,4 +187,3 @@
     window.ctx.setRepoId("@{ RequestContext.getRepository.persistedId }")
 </script>
 
-<script type="module" src="/static/js/folders/dragon-drop.js"></script>
